Refactor: move form init code and logic into separate file
