+++ app/javascript/administrate/global_search/index.js
@@ -1,8 +1,6 @@
 import tablesort from 'tablesort';
-import tippy from 'tippy.js';
-import 'tippy.js/dist/tippy.css';
-import 'tippy.js/themes/light.css';
-import 'tippy.js/animations/shift-toward.css';
+
+import initGlobalSearchForm from 'administrate/global_search/form';
 
 const GLOBAL_SEARCH_URL = `/assets/global_search`;
 const USER_VIEW_URL = id => `/admin/users/${id}`;
@@ -10,6 +8,12 @@ const ASSET_VIEW_URL = id => `/admin/assets/${id}`;
 const LOCATION_VIEW_URL = id => `/admin/locations/${id}`;
 
 const initSearchIndex = () => {
+  const $results = $('#search-results__body');
+  const $welcomeScreen = $('#welcome-screen');
+  const $noResultsScreen = $('#no-results-screen');
+
+  const { getFormState } = initGlobalSearchForm(submitSearch);
+
   // Table sorting & related styling
   const tableElem = document.getElementById('search-results');
   const tableSorter = tablesort(tableElem);
@@ -36,98 +40,9 @@ const initSearchIndex = () => {
     `);
   });
 
-  const $results = $('#search-results__body');
-  const $welcomeScreen = $('#welcome-screen');
-  const $noResultsScreen = $('#no-results-screen');
-
-  // Form setup
-  const $searchForm = $('.global-search__form');
-  const $nameQueryInput = $('.global-search__input');
-  const $nameQueryClearBtn = $('.global-search__form__clear-btn');
-
-  $nameQueryClearBtn.on('click', e => {
-    e.preventDefault();
-    $nameQueryInput.val('');
-    submitSearch();
-  });
-
-  $searchForm.on('submit', e => {
-    e.preventDefault();
-    submitSearch();
-  });
-
-  // Filter button setup
-  const $tagFilterBtn = $('#tag-filter-btn');
-  const $userFilterBtn = $('#user-filter-btn');
-  const $locationFilterBtn = $('#location-filter-btn');
-  const $checkedInFilterBtn = $('#checked-in-filter-btn');
-  const $checkedOutFilterBtn = $('#checked-out-filter-btn');
-  const $lateFilterBtn = $('#late-filter-btn');
-
-  const $tagTooltip = $('#tag-tooltip');
-  const $userTooltip = $('#user-tooltip');
-  const $locationTooltip = $('#location-tooltip');
-
-  const tippyOptions = {
-    allowHTML: true,
-    trigger: 'click',
-    placement: 'bottom',
-    interactive: true,
-    animation: 'shift-toward',
-    theme: 'light',
-    appendTo: document.body,
-  };
-
-  tippy($tagFilterBtn[0], {
-    content: $tagTooltip.html(),
-    ...tippyOptions,
-  });
-  tippy($userFilterBtn[0], {
-    content: $userTooltip.html(),
-    ...tippyOptions,
-  });
-  tippy($locationFilterBtn[0], {
-    content: $locationTooltip.html(),
-    ...tippyOptions,
-  });
-
-  // TODO - selectize dropdowns
-
-  // Restore any previous search from current session
-  restoreSearch();
-
-  // Search query + filters represented as JSON
-  function getFormState() {
-    return {
-      query: $nameQueryInput.val(),
-      filters: {},
-    };
-  }
-
-  function saveSearch() {
-    window.sessionStorage.setItem(
-      'global-search-state',
-      JSON.stringify(getFormState())
-    );
-  }
-
-  function restoreSearch() {
-    const saved = window.sessionStorage.getItem('global-search-state');
-    if (typeof saved === 'string') {
-      const { query /* , filters */ } = JSON.parse(saved);
-
-      $nameQueryInput.val(query);
-
-      // TODO: restore filter state
-
-      submitSearch();
-    }
-  }
-
-  function submitSearch() {
-    saveSearch();
+  function submitSearch(formState) {
     clearResults();
-    $.post(GLOBAL_SEARCH_URL, getFormState())
+    $.post(GLOBAL_SEARCH_URL, formState)
       .then(results => {
         renderPage(!!results);
         if (results) {
